+++ src/store/index.js
@@ -1,10 +1,9 @@
 import Vue from 'vue';
 import Vuex from 'vuex';
 import Auth from './modules/auth';
-import Lodging from './modules/Lodging';
+import Lodging from './modules/lodging/';
 import Maintenance from './modules/Maintenance';
 import Payments from './modules/Payments';
-import Period from './modules/Period';
 import Persons from './modules/Persons';
 import Place from './modules/Place';
 import Room from './modules/rooms';
@@ -18,7 +17,6 @@ export default new Vuex.Store({
 		Lodging,
 		Maintenance,
 		Payments,
-		Period,
 		Persons,
 		Place,
 		Room,
