+++ src/store/modules/lodging/getters.js
@@ -1,0 +1,20 @@
+const getters = {
+	bottomSheet: state => state.bottomSheet,
+	message: state => state.message,
+	lodgingPersons: state => state.lodgingPersons,
+	updatingService: state => state.updatingService,
+	mirrorLodging: state => state.mirrorLodging,
+	lodgingSelect: state => state.lodgingSelect,
+	loading: state => state.loading,
+	editMode: state => state.editMode,
+	lodgings: state => state.lodgings,
+	rangeDatePayments: state => state.rangeDatePayments,
+	lodgingsPlace: state => state.lodgingsPlace,
+	countLogingsPlace: state => state.countLogingsPlace,
+	rangeDate: state => state.rangeDate,
+	periods: state => state.periods,
+	places: state => state.Places,
+	place: state => state.place,
+};
+
+export default getters;
